PALLADIO-498 Use new StoEx parser in PCMRandomVariable
