+++ bundles/org.palladiosimulator.pcm/META-INF/MANIFEST.MF
@@ -81,5 +81,6 @@ Require-Bundle: org.eclipse.core.runtime,
  de.uka.ipd.sdq.units;visibility:=reexport,
  de.uka.ipd.sdq.errorhandling,
  org.eclipse.ocl.ecore;bundle-version="1.2.1",
- org.palladiosimulator.pcm.workflow;resolution:=optional
+ org.palladiosimulator.pcm.workflow;resolution:=optional,
+ org.palladiosimulator.commons.stoex.api;bundle-version="4.3.0"
 Bundle-ActivationPolicy: lazy
