+++ tests/org.palladiosimulator.pcm.tests/META-INF/MANIFEST.MF
@@ -6,6 +6,9 @@ Bundle-Version: 4.3.0.qualifier
 Fragment-Host: org.palladiosimulator.pcm;bundle-version="4.3.0"
 Automatic-Module-Name: org.palladiosimulator.pcm.tests
 Bundle-RequiredExecutionEnvironment: JavaSE-11
-Require-Bundle: tools.mdsd.library.standalone.initialization;bundle-version="0.1.0"
+Require-Bundle: tools.mdsd.library.standalone.initialization;bundle-version="0.1.0",
+ org.hamcrest.core;bundle-version="1.3.0",
+ org.hamcrest.library;bundle-version="1.3.0",
+ org.palladiosimulator.commons.stoex.ui;bundle-version="4.3.0"
 Import-Package: org.junit;version="4.12.0",
  org.junit.jupiter.api;version="5.6.0"
